[Android] Added new view for new tab page This is is to avoid null pointer exception
Chromium change:
https://chromium.googlesource.com/chromium/src/+/efb886565e933f36643147e9c9ee7ce19dc17bce

Cryptids: Injecting view into NTP logo space

This adds a GIF view that will eventually be used for cryptid rendering.

The general approach is:
* The existing logo is wrapped in a FrameLayout with a new ViewStub.
* When we get a callback indicating that no doodle is present, the NTP
  requests a drawable from the PCR and injects it into an ImageView
  inflated from the ViewStub.

Out-of-scope for this CL is to actually populate the GIF image data; I
have been testing locally with a res/raw file read into a byte array.

Change-Id: Ib5deae5322856c9d8e8a68f0fd46cdb2aaadbe79
Bug: 1061949